+++ src/Crypt.php
@@ -8,9 +8,9 @@
 namespace CryptForWordPress;
 
 // prevent direct access.
-use WP_Error;
-
 defined( 'ABSPATH' ) || exit;
+
+use WP_Error;
 
 /**
  * Object to handle crypt tasks.
@@ -74,6 +74,13 @@ class Crypt {
 		// get the place object.
 		$place_obj = $this->get_place();
 		if ( ! $place_obj instanceof Place_Base ) {
+			// log this error.
+			$this->add_error(
+				'save_place_not_available',
+				'Could not find any place to save the key for encryption.'
+			);
+
+			// do nothing more.
 			return false;
 		}
 		$place_obj->load();
@@ -113,6 +120,13 @@ class Crypt {
 
 		// bail if the method could not be found.
 		if ( false === $method_obj ) {
+			// log this error.
+			$this->add_error(
+				'no_method_available',
+				'Could not find any supported method to encrypt strings.'
+			);
+
+			// do nothing more.
 			return '';
 		}
 
@@ -133,6 +147,13 @@ class Crypt {
 
 		// bail if the method could not be found.
 		if ( false === $method_obj ) {
+			// log this error.
+			$this->add_error(
+				'no_method_available',
+				'Could not find any supported method to encrypt strings.'
+			);
+
+			// do nothing more.
 			return '';
 		}
 
@@ -250,7 +271,10 @@ class Crypt {
 	 * Set the plugin file.
 	 *
 	 * @param string $plugin_file The absolute path to the plugin file.
+	 *
 	 * @return void
+	 * @noinspection PhpUnused
+	 * @noinspection PhpUnused
 	 */
 	public function set_plugin_file( string $plugin_file ): void {
 		$this->plugin_file = $plugin_file;
@@ -322,7 +346,7 @@ class Crypt {
 			return array();
 		}
 
-		// bail if configuration is not an array.
+		// bail if the configuration is not an array.
 		if ( ! is_array( $this->configuration[ $method_name ] ) ) {
 			return array();
 		}
@@ -341,7 +365,7 @@ class Crypt {
 	}
 
 	/**
-	 * Set custom configuration for this object.
+	 * Set the custom configuration for this object.
 	 *
 	 * @param array<string,array<string,mixed>|string> $configurations List of configurations.
 	 * @return void
@@ -359,9 +383,11 @@ class Crypt {
 		$places = array(
 			'CryptForWordPress\Places\WpConfig',
 			'CryptForWordPress\Places\MuPlugin',
+			'CryptForWordPress\Places\Database',
 			'CryptForWordPress\Places\CustomFile',
 			'CryptForWordPress\Places\EnvironmentVariable',
 			'CryptForWordPress\Places\ServerVariable',
+			'CryptForWordPress\Places\WordPressSalts',
 		);
 
 		/**
@@ -416,7 +442,7 @@ class Crypt {
 	}
 
 	/**
-	 * Return the place where the token should be saved.
+	 * Return the place, where the token should be saved.
 	 *
 	 * @access private
 	 *
@@ -429,6 +455,9 @@ class Crypt {
 			if ( ! $obj->is_usable() ) {
 				continue;
 			}
+
+			// set the configuration.
+			$obj->set_config( $this->get_config() );
 
 			// return this place object.
 			return $obj;
@@ -451,6 +480,13 @@ class Crypt {
 
 		// bail if no place could be loaded.
 		if ( ! $place_obj instanceof Place_Base ) {
+			// log this error.
+			$this->add_error(
+				'save_place_not_available',
+				'Could not find any place to save the key for encryption.'
+			);
+
+			// do nothing more.
 			return;
 		}
 
@@ -490,7 +526,7 @@ class Crypt {
 	}
 
 	/**
-	 * Add an error the list.
+	 * Add an error to the list.
 	 *
 	 * @param string              $code The error code.
 	 * @param string              $message The error message.
@@ -533,7 +569,7 @@ class Crypt {
 	}
 
 	/**
-	 * Return whether errors has been occurred.
+	 * Return whether errors have been occurred.
 	 *
 	 * @return bool
 	 */
@@ -546,6 +582,8 @@ class Crypt {
 	 * Reset the list of errors.
 	 *
 	 * @return void
+	 * @noinspection PhpUnused
+	 * @noinspection PhpUnused
 	 */
 	public function clear_errors(): void {
 		$this->errors = null;